A PowerShell port of the testing script "Malx" used on "The PC Security Channel"
A PowerShell port of the testing script "Malx" used on "The PC Security Channel". Meant to work locally with "theZoo" malware repository. The script will extract the malware from each respective zip folder using the "7Zip4Powershell" module. The module will be installed if it is not already. This script is meant to automate the process of testing a AV/EDR/XDR/SEIM tool against different malware samples.